+++ minicube-generation/CHANGES.md
@@ -1,4 +1,4 @@
-## Changes in 1.0.1
+## Changes in 1.1
 * Changed cube generation process to start processes until maximum is reached
 * Use `four_d` parameter for sentinelhub store to decrease number of requests
 * Introduced `COMPONONENTVERSIONS.md`